+++ packages/abi-ts/.npmignore
@@ -1,0 +1,9 @@
+*
+
+!abi/**
+!src/**
+!ts/**
+!types/**
+!package.json
+!README.md
+!dist/**
