+++ packages/abi-ts/README.md
@@ -1,0 +1,7 @@
+# ABI -> TS
+
+Convert a directory of JSON ABI files to a directory of TS files with `as const`.
+
+```console
+pnpm abi-ts --input abi --output abi-ts
+```
