+++ packages/abi-ts/package.json
@@ -1,0 +1,49 @@
+{
+  "name": "@latticexyz/abi-ts",
+  "version": "2.0.0-next.5",
+  "description": "Convert a directory of JSON ABI files to a directory of TS files with `as const`",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/latticexyz/mud.git",
+    "directory": "packages/abi-ts"
+  },
+  "license": "MIT",
+  "type": "module",
+  "exports": {
+    ".": "./dist/index.js"
+  },
+  "typesVersions": {
+    "*": {
+      "index": [
+        "./src/index.ts"
+      ]
+    }
+  },
+  "bin": {
+    "abi-ts": "./dist/abi-ts.js"
+  },
+  "scripts": {
+    "build": "tsup",
+    "clean": "rimraf dist",
+    "dev": "tsup --watch",
+    "test": "vitest typecheck --run --passWithNoTests && vitest --run --passWithNoTests",
+    "test:ci": "pnpm run test"
+  },
+  "dependencies": {
+    "chalk": "^5.3.0",
+    "execa": "^7.0.0",
+    "glob": "^8.0.3",
+    "yargs": "^17.7.1"
+  },
+  "devDependencies": {
+    "@types/glob": "^7.2.0",
+    "@types/node": "^18.15.11",
+    "@types/yargs": "^17.0.10",
+    "tsup": "^6.7.0",
+    "vitest": "0.31.4"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "gitHead": "914a1e0ae4a573d685841ca2ea921435057deb8f"
+}
