+++ packages/store/package.json
@@ -39,7 +39,7 @@
     "build:js": "tsup",
     "build:mud": "tsx ./ts/scripts/tablegen.ts",
     "build:tightcoder": "tsx ./ts/scripts/generate-tightcoder.ts",
-    "build:ts": "mud abi-ts --input abi --output abi-ts && prettier --write 'abi-ts/**/*.ts'",
+    "build:ts": "abi-ts --input abi --output abi-ts && prettier --write 'abi-ts/**/*.ts'",
     "clean": "pnpm run clean:abi && pnpm run clean:js && pnpm run clean:mud && pnpm run clean:ts",
     "clean:abi": "rimraf abi",
     "clean:js": "rimraf dist",
@@ -60,7 +60,7 @@
     "zod": "^3.21.4"
   },
   "devDependencies": {
-    "@latticexyz/cli": "workspace:*",
+    "@latticexyz/abi-ts": "workspace:*",
     "@types/ejs": "^3.1.1",
     "@types/mocha": "^9.1.1",
     "@types/node": "^18.15.11",
