+++ packages/world/package.json
@@ -35,7 +35,7 @@
     "build:abi": "forge build --extra-output-files abi --out abi --skip test script",
     "build:js": "tsup",
     "build:mud": "tsx ./ts/scripts/tablegen.ts && tsx ./ts/scripts/worldgen.ts",
-    "build:ts": "mud abi-ts --input abi --output abi-ts && prettier --write 'abi-ts/**/*.ts'",
+    "build:ts": "abi-ts --input abi --output abi-ts && prettier --write 'abi-ts/**/*.ts'",
     "clean": "pnpm run clean:abi && pnpm run clean:js && pnpm run clean:mud && pnpm run clean:ts",
     "clean:abi": "rimraf abi",
     "clean:js": "rimraf dist",
@@ -56,7 +56,7 @@
     "zod": "^3.21.4"
   },
   "devDependencies": {
-    "@latticexyz/cli": "workspace:*",
+    "@latticexyz/abi-ts": "workspace:*",
     "@types/ejs": "^3.1.1",
     "@types/glob": "^7.2.0",
     "@types/mocha": "^9.1.1",
